feat: api endpoint to get from bucket and gateway using worker binding
