+++ packages/api/src/perma-cache/get.js
@@ -1,96 +1,28 @@
 /* eslint-env serviceworker, browser */
 /* global Response */
 
-import { JSONResponse } from '../utils/json-response.js'
+import { getSourceUrl, getNormalizedUrl } from '../utils/url.js'
+import { UrlNotFoundError } from '../errors.js'
 
 /**
  * @typedef {import('../env').Env} Env
  */
 
 /**
- * Handle perma-cache put request
+ * Handle perma-cache get request
  *
  * @param {Request} request
  * @param {Env} env
  */
-export async function permaCacheListGet(request, env) {
-  const requestUrl = new URL(request.url)
-  const { searchParams } = requestUrl
-  const { size, page, sort, order } = parseSearchParams(searchParams)
+export async function permaCacheGet(request, env) {
+  const sourceUrl = getSourceUrl(request, env)
+  const normalizedUrl = getNormalizedUrl(sourceUrl, env)
+  const r2Key = normalizedUrl.toString()
 
-  const entries = await env.db.listPermaCache(request.auth.user.id, {
-    size,
-    page,
-    sort,
-    order,
-  })
-
-  // Get next page link
-  const headers =
-    entries.length === size
-      ? {
-          Link: `<${requestUrl.pathname}?size=${size}&page=${
-            page + 1
-          }>; rel="next"`,
-        }
-      : undefined
-  return new JSONResponse(entries, { headers })
-}
-
-/**
- * @param {URLSearchParams} searchParams
- */
-function parseSearchParams(searchParams) {
-  // Parse size parameter
-  let size = 25
-  if (searchParams.has('size')) {
-    const parsedSize = parseInt(searchParams.get('size'))
-    if (isNaN(parsedSize) || parsedSize <= 0 || parsedSize > 1000) {
-      throw Object.assign(new Error('invalid page size'), { status: 400 })
-    }
-    size = parsedSize
+  const r2Object = await env.SUPERHOT.get(r2Key)
+  if (r2Object) {
+    return new Response(r2Object.body)
   }
 
-  // Parse cursor parameter
-  let page = 0
-  if (searchParams.has('page')) {
-    const parsedPage = parseInt(searchParams.get('page'))
-    if (isNaN(parsedPage) || parsedPage <= 0) {
-      throw Object.assign(new Error('invalid page number'), { status: 400 })
-    }
-    page = parsedPage
-  }
-
-  // Parse sort parameter
-  let sort = 'date'
-  if (searchParams.has('sort')) {
-    const parsedSort = searchParams.get('sort')
-    if (parsedSort !== 'date' && parsedSort !== 'size') {
-      throw Object.assign(
-        new Error('invalid list sort, either "date" or "size"'),
-        { status: 400 }
-      )
-    }
-    sort = parsedSort
-  }
-
-  // Parse order parameter
-  let order = 'asc'
-  if (searchParams.has('order')) {
-    const parsedOrder = searchParams.get('order')
-    if (parsedOrder !== 'asc' && parsedOrder !== 'desc') {
-      throw Object.assign(
-        new Error('invalid list sort order, either "asc" or "desc"'),
-        { status: 400 }
-      )
-    }
-    sort = parsedOrder
-  }
-
-  return {
-    size,
-    page,
-    sort,
-    order,
-  }
+  throw new UrlNotFoundError()
 }
