+++ packages/api/src/perma-cache/index.js
@@ -1,4 +1,5 @@
 export { permaCachePost } from './post.js'
-export { permaCacheListGet } from './get.js'
+export { permaCacheGet } from './get.js'
 export { permaCacheAccountGet } from './account.js'
+export { permaCacheListGet } from './list.js'
 export { permaCacheDelete } from './delete.js'
