+++ packages/edge-gateway/test/utils.js
@@ -1,4 +1,3 @@
-import { concat } from 'uint8arrays/concat'
 import { Miniflare } from 'miniflare'
 
 export function getMiniflare() {
@@ -14,8 +13,14 @@ export function getMiniflare() {
     buildCommand: undefined,
     wranglerConfigEnv: 'test',
     modules: true,
-    bindings: {
-      SUPERHOT: createR2Bucket(),
+    mounts: {
+      api: {
+        scriptPath: './test/scripts/api.js',
+        modules: true,
+      },
+    },
+    serviceBindings: {
+      API: 'api',
     },
   })
 }
@@ -30,63 +35,4 @@ export function getGatewayRateLimitsName() {
 
 export function getSummaryMetricsName() {
   return 'SUMMARYMETRICS'
-}
-
-function createR2Bucket() {
-  const bucket = new Map()
-
-  return {
-    put: async (key, value, putOpts = {}) => {
-      let data = new Uint8Array([])
-      const reader = value.getReader()
-      try {
-        while (true) {
-          const { done, value } = await reader.read()
-          if (done) {
-            break
-          }
-          data = concat([data, value])
-        }
-      } finally {
-        reader.releaseLock()
-      }
-
-      bucket.set(key, {
-        body: data,
-        httpMetadata: putOpts.httpMetadata || {},
-        customMetadata: putOpts.customMetadata || {},
-      })
-
-      return Promise.resolve({
-        httpMetadata: putOpts.httpMetadata,
-        customMetadata: putOpts.customMetadata,
-      })
-    },
-    get: async (key) => {
-      const value = bucket.get(key)
-      if (!value) {
-        return undefined
-      }
-
-      const response = new Response(value.body, { status: 200 })
-
-      return Promise.resolve(
-        Object.assign(response, {
-          httpMetadata: value.httpMetadata || {},
-          customMetadata: value.customMetadata || {},
-        })
-      )
-    },
-    head: async (key) => {
-      const value = bucket.get(key)
-      if (!value) {
-        return undefined
-      }
-
-      return Promise.resolve({
-        httpMetadata: value.httpMetadata || {},
-        customMetadata: value.customMetadata || {},
-      })
-    },
-  }
 }
