+++ packages/edge-gateway/wrangler.toml
@@ -32,6 +32,7 @@ kv_namespaces = [{ binding = "DENYLIST", id = "785cf627e913468ca5319523ae929def"
 [env.production.vars]
 IPFS_GATEWAYS = "[\"https://ipfs.io\", \"https://cf.nftstorage.link\", \"https://pinata.nftstorage.link\"]"
 GATEWAY_HOSTNAME = 'ipfs.nftstorage.link'
+EDGE_GATEWAY_API_URL = 'https://api.nftstorage.link'
 DATABASE_URL = "https://nft-storage-pgrest-prod.herokuapp.com"
 DEBUG = "false"
 ENV = "production"
@@ -44,9 +45,11 @@ bindings = [
   {name = "GATEWAYREDIRECTCOUNTER", class_name = "GatewayRedirectCounter0"}
 ]
 
-[[env.production.r2_buckets]]
-bucket_name = "super-hot"
-binding = "SUPERHOT"
+[[env.production.services]]
+binding = "API"
+type = "service"
+service = "nftstorage-link-api-production"
+environment = "production"
 
 # Staging!
 [env.staging]
@@ -64,6 +67,7 @@ kv_namespaces = [{ binding = "DENYLIST", id = "f4eb0eca32e14e28b643604a82e00cb3"
 [env.staging.vars]
 IPFS_GATEWAYS = "[\"https://ipfs.io\", \"https://cf.nftstorage.link\", \"https://pinata.nftstorage.link\"]"
 GATEWAY_HOSTNAME = 'ipfs-staging.nftstorage.link'
+EDGE_GATEWAY_API_URL = 'https://api.nftstorage.link'
 DATABASE_URL = "https://nft-storage-pgrest-staging.herokuapp.com"
 DEBUG = "true"
 ENV = "staging"
@@ -76,9 +80,11 @@ bindings = [
   {name = "GATEWAYREDIRECTCOUNTER", class_name = "GatewayRedirectCounter0"}
 ]
 
-[[env.staging.r2_buckets]]
-bucket_name = "super-hot-staging"
-binding = "SUPERHOT"
+[[env.staging.services]]
+binding = "API"
+type = "service"
+service = "nftstorage-link-api-staging"
+environment = "production"
 
 # Test!
 [env.test]
@@ -88,6 +94,7 @@ kv_namespaces = [{ binding = "DENYLIST" }]
 [env.test.vars]
 IPFS_GATEWAYS = "[\"http://127.0.0.1:9081\", \"http://localhost:9082\", \"http://localhost:9083\"]"
 GATEWAY_HOSTNAME = 'ipfs.localhost:8787'
+EDGE_GATEWAY_API_URL = 'http://localhost:8787'
 DEBUG = "true"
 ENV = "test"
 
@@ -108,6 +115,7 @@ kv_namespaces = [{ binding = "DENYLIST", id = "" }]
 [env.vsantos.vars]
 GATEWAY_HOSTNAME = 'ipfs.localhost:8787'
 IPFS_GATEWAYS = "[\"https://ipfs.io\"]"
+EDGE_GATEWAY_API_URL = 'http://localhost:8787'
 
 [env.vsantos.durable_objects]
 bindings = [
